pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testgen-ai"
version = "0.1.0"
description = "The Autonomous QA Agent from Your CLI - AI-powered test generation and execution"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "TBD"}
authors = [
    {name = "Jay Patil", email = "your.email@example.com"}
]
keywords = [
    "testing",
    "qa",
    "ai",
    "llm",
    "test-generation",
    "automation",
    "pytest",
    "tdd"
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "watchdog>=3.0.0",
    "pytest>=7.0.0",
    "pytest-json-report>=1.5.0",
    "playwright>=1.40.0",
    "jinja2>=3.1.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/JayPatil165/TestGen-AI"
Repository = "https://github.com/JayPatil165/TestGen-AI"
Issues = "https://github.com/JayPatil165/TestGen-AI/issues"
Documentation = "https://github.com/JayPatil165/TestGen-AI/tree/main/docs"

[project.scripts]
testgen = "testgen.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
testgen = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]
